Completed character search + modal
Wanted to reduce requests as much as possible when searching and using only options after a search.
Tried a lot of solution such as modifying apollo cache, creating a client store,
and creating reactive variables to store a local version of the results. It's necessary due to my search algorithm.
For example, ni hao, it will search for both ni and ni hao. If a user searched ni hao and then ni.
It is not necessary to request again, but apollo fetches every time a variable is changed.
Decided to just store the query, and if only options are changed after a search,
it will use the same query results and only filter current results.